Introduce close/1, rework socket activation.
close/1 is easy to implement. Basically add a state for a closed socket, and then use
the new state to reply back correctly when someone wants to communicate on the
socket. Also, when a socket closes, remove its internal reference to the socket,
making the code crash on any attempt to use the socket afterwards.

Socket activation is reworked, so the FSM now has to explicitly ask for a new activation
of a received message. This in effect makes sure that we can implement the recv/1
interface on the socket in question. Next up is buffering of sockets, which has to be
correctly handled as we go along.