Use the cookie packet for peer information only.
The code assumed we already had access to a peer public key, EC. But this public key
is packed up inside the cookie and the application is explicitly made such that an enemy
can't store state in our end. A future version of the application might delay process generation
for instance.

Once we have checked the vouch packet, however, store the peer_public_key into the
state of the process for future reference.