Improve code flow for packet reception, eliminate activate_socket/2.
Handling a TCP packet is now controlled by a tail-call sequence which returns
{next_state, …} in the end. This is easier since there is no reason for the reader
to understand the next | hold atom alternation. Every time we say 'next', we
basically tell the socket to activate once.

The `activate_socket/2` function is not useful anymore. It is a simple function,
and it just calls `inet:setopts/2` with `{active, once}`. Rather than forcing the
reader to understand another function, we just inline the setopts call directly
where it is used.