Implement metadata in vouch and new ready message
This change adds a new message ‘ready’ from the
server side in response to the vouch message.

Both the ‘vouch’ and ‘ready’ messages carry a
key/value set, that is user defined.  The idea
is that this can carry information pertaining to
the content of the messages; such as which protocol
is to be used, or additional information about why
a peer should be trusted.

This change draws inspiration from CurveZMQ, which
provides a similar facility.