Add curve_tun:metadata/1 API and option
This change adds a new API function to inspect
the peer’s metadata; which is available after
successfully returning from connect and/or accept.

Metadata to send to the peer is configured as
a option to connect/listen.

This change also adds some basic tests to the
functionality.

Design: Metadata is passed as binaries, not atoms,
because atoms are not garbage collected so a peer
can flood the atom table.